refactor: extracted chat logic into a utility function
