+++ pages/index.tsx
@@ -2,120 +2,15 @@ import Layout from '@/components/layout';
 import LoadingDots from '@/components/ui/LoadingDots';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/components/ui/accordion';
 import styles from '@/styles/Home.module.css';
-import { Message } from '@/types/chat';
-import { Document } from 'langchain/document';
+import useChat from '@/utils/use-chat';
 import { useSession } from 'next-auth/react';
 import Image from 'next/image';
-import { useEffect, useRef, useState } from 'react';
 import ReactMarkdown from 'react-markdown';
 import TypewriterComponent from 'typewriter-effect';
 
 export default function Home() {
-  const [query, setQuery] = useState<string>('');
-  const [loading, setLoading] = useState<boolean>(false);
-  const [error, setError] = useState<string | null>(null);
-  const [messageState, setMessageState] = useState<{
-    messages: Message[];
-    pending?: string;
-    history: [string, string][];
-    pendingSourceDocs?: Document[];
-  }>({
-    messages: [
-      {
-        message: 'Hi, what you like to learn from the available documents?',
-        type: 'apiMessage',
-      },
-    ],
-    history: [],
-  });
-
-  const { messages, history } = messageState;
-  const messageListRef = useRef<HTMLDivElement>(null);
-  const textAreaRef = useRef<HTMLTextAreaElement>(null);
+  const { loading, error, messages, textAreaRef, messageListRef, query, setQuery, handleEnter, handleSubmit } = useChat();
   const { data: session } = useSession();
-
-  useEffect(() => {
-    textAreaRef.current?.focus();
-  }, []);
-
-  //handle form submission
-  async function handleSubmit(e: any) {
-    e.preventDefault();
-
-    setError(null);
-
-    if (!query) {
-      alert('Please input a question');
-      return;
-    }
-
-    const question = query.trim();
-
-    setMessageState((state) => ({
-      ...state,
-      messages: [
-        ...state.messages,
-        {
-          type: 'userMessage',
-          message: question,
-        },
-      ],
-    }));
-
-    setLoading(true);
-    setQuery('');
-
-    try {
-      const response = await fetch('/api/chat', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          question,
-          history,
-        }),
-      });
-      const data = await response.json();
-      console.log('data', data);
-
-      if (data.error) {
-        setError(data.error);
-      } else {
-        setMessageState((state) => ({
-          ...state,
-          messages: [
-            ...state.messages,
-            {
-              type: 'apiMessage',
-              message: data.text,
-              sourceDocs: data.sourceDocuments,
-            },
-          ],
-          history: [...state.history, [question, data.text]],
-        }));
-      }
-      console.log('messageState', messageState);
-
-      setLoading(false);
-
-      //scroll to bottom
-      messageListRef.current?.scrollTo(0, messageListRef.current.scrollHeight);
-    } catch (error) {
-      setLoading(false);
-      setError('An error occurred while fetching the data. Please try again.');
-      console.log('error', error);
-    }
-  }
-
-  //prevent empty submissions
-  const handleEnter = (e: any) => {
-    if (e.key === 'Enter' && query) {
-      handleSubmit(e);
-    } else if (e.key == 'Enter') {
-      e.preventDefault();
-    }
-  };
 
   return (
     <>
@@ -270,7 +165,6 @@ export default function Home() {
           </main>
         </div>
         <footer className="m-auto p-4">
-          <FileListButton/>
           <a href="https://twitter.com/murithi_myke">
             Powered by LangChainAI. Current version built by (Twitter: @murithi_myke).
           </a>
